requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
cffi==1.14.0
click==7.1.1
cryptography==2.9
flasgger==0.9.4
Flask==1.1.2
Flask-HTTPAuth==3.3.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
mistune==0.8.4
pathspec==0.8.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.16.0
python-dotenv==0.12.0
PyYAML==5.3.1
regex==2020.4.4
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.1
zipp==3.1.0
